# api/device-service.openapi.yaml
openapi: 3.0.3
info:
  title: Device Service API
  description: |
    REST API микросервиса device-service (домен «Управление устройствами»)
    экосистемы «Тёплый дом». Синхронные операции: регистрация устройства
    в режиме самообслуживания, получение информации, смена состояния,
    отправка команд (асинхронное исполнение через брокер) и чтение истории телеметрии.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Локальный запуск через API Gateway
security:
  - bearerAuth: []
paths:
  /devices:
    post:
      operationId: registerDevice
      summary: Регистрация устройства (self-service)
      description: Пользователь самостоятельно регистрирует купленное устройство в своём доме.
      tags: [devices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
            examples:
              default:
                summary: Регистрация датчика температуры
                value:
                  serial_number: "TH-000123456"
                  type_id: 1
                  house_id: 42
                  name: "Датчик температуры (гостиная)"
      responses:
        '201':
          description: Устройство зарегистрировано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
              examples:
                default:
                  value:
                    id: 1001
                    serial_number: "TH-000123456"
                    type_id: 1
                    house_id: 42
                    name: "Датчик температуры (гостиная)"
                    status: "off"
                    firmware_version: "1.0.0"
                    last_seen_at: null
                    created_at: "2025-07-14T10:15:00Z"
        '400':
          description: Некорректное тело запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                default:
                  value:
                    error: "field 'serial_number' is required"
        '409':
          description: Устройство с таким serial_number уже зарегистрировано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                default:
                  value:
                    error: "device with serial_number 'TH-000123456' already exists"
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                default:
                  value:
                    error: "internal server error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-токен, выдаётся API Gateway после аутентификации пользователя.
...
  schemas:
    DeviceCreate:
      type: object
      required: [serial_number, type_id, house_id, name]
      properties:
        serial_number:
          type: string
          example: "TH-000123456"
        type_id:
          type: integer
          format: int64
          example: 1
        house_id:
          type: integer
          format: int64
          example: 42
        name:
          type: string
          example: "Датчик температуры (гостиная)"
...
    Device:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1001
        serial_number:
          type: string
          example: "TH-000123456"
        type_id:
          type: integer
          format: int64
          example: 1
        house_id:
          type: integer
          format: int64
          example: 42
        name:
          type: string
          example: "Датчик температуры (гостиная)"
        status:
          type: string
          enum: [on, off]
          example: "on"
        firmware_version:
          type: string
          nullable: true
          example: "1.0.0"
        last_seen_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-07-14T10:20:00Z"
        created_at:
          type: string
          format: date-time
          example: "2025-07-14T10:15:00Z"
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "device 1001 not found"